dependabot updates as of 2023-05-02
I am trying to find the most effective way of managing multiple `dependabot` PRs, both in terms of my time and that of others.

It appears that doing a "rollup-PR" (such as this one) into `dev` (after testing locally) is good for a deployed test, and, if all is well, then the `dependeabot` can be manually approved. This avoids having to interrupt someone else to deal with these low-level interrupts.

Let's try this for a few weeks and see how it goes.