Brave Talk Marvin to staging
Brave folks: see https://docs.google.com/document/d/1Cb15YelULNfjM50eanx6asdn4F354ufxkTPvtpNSZDM for the specification and https://github.com/brave-intl/brave-talk-server/pull/50 for the changes to the Brave Talk server

Briefly: a premium moderator may turn on transcriptions (speech-to-text) in addition to A/V recording. As with A/V recordings, the moderator has access to a transcription file that "lives" for 24h (with links on the home page). Unlike A/V recordings, all participants receive incremental transcription "chunks", which are visible to "Brave Leo", thereby allowing the participants to (1) ask what happened before they joined and (2) ask Leo questions related to the call.